# shared/openapi.yaml
openapi: 3.1.0
info:
  title: Icarus API
  description: Contract v1 for the Icarus server (shared/api-contract.md). Times are RFC 3339 UTC with Z, except fields ending in _ms (epoch milliseconds). Metrics are estimates, not medical readings.
  version: 1.0.0
paths:
...
  /v1/alarms:
    get:
      tags:
      - Alarms
      operationId: list_alarms
      responses:
        '200':
          description: Alarms that are not deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlarmList'
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - session: []
      - device: []
    post:
      tags:
      - Alarms
      operationId: create_alarm
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlarmCreate'
        required: true
      responses:
        '201':
          description: The created alarm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alarm'
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - session: []
      - device: []
...
  /v1/alarms/{id}:
    delete:
      tags:
      - Alarms
      summary: 'Soft delete: the row stays as a tombstone with a new version, so the app learns of the deletion.'
      operationId: delete_alarm
      parameters:
      - name: id
        in: path
        description: Id of the resource.
        required: true
        schema:
          type: string
          format: uuid
      - name: If-Match
        in: header
        description: Optional. When sent, a stale value gets 409.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      responses:
        '204':
          description: Soft-deleted. The version is bumped so the app sees a tombstone.
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - session: []
      - device: []
    patch:
      tags:
      - Alarms
      operationId: patch_alarm
      parameters:
      - name: id
        in: path
        description: Id of the resource.
        required: true
        schema:
          type: string
          format: uuid
      - name: If-Match
        in: header
        description: Current version. Required; a stale value gets 409.
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AlarmPatch'
        required: true
      responses:
        '200':
          description: The updated alarm.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alarm'
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - session: []
      - device: []
  /v1/alarms/{id}/test:
    post:
      tags:
      - Alarms
      summary: |-
        Queues a test dispatch with the alarm's own rhythm and channels. The app and dispatcher treat it
        like a webhook dispatch.
      operationId: test_alarm
      parameters:
      - name: id
        in: path
        description: Id of the resource.
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '202':
          description: Queued with the alarm rhythm and channels.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DispatchIdResponse'
        default:
          description: An error as application/problem+json (api-contract.md Errors).
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      security:
      - session: []
      - device: []
...
components:
  schemas:
...
    Alarm:
      type: object
      description: '`Alarm` in api-contract.md.'
      required:
      - id
      - kind
      - label
      - rhythm
      - channels
      - enabled
      - version
      - updated_at
      properties:
        channels:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
        deleted_at:
          type:
          - string
          - 'null'
          format: date-time
        enabled:
          type: boolean
        id:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/AlarmKind'
        label:
          type: string
        rhythm:
          $ref: '#/components/schemas/Rhythm'
        schedule:
          oneOf:
          - $ref: '#/components/schemas/Schedule'
          - type: 'null'
        updated_at:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
    AlarmCreate:
      type: object
      required:
      - kind
      - label
      - rhythm
      - channels
      properties:
        channels:
          type: array
          items:
            $ref: '#/components/schemas/Channel'
        enabled:
          type: boolean
        id:
          type:
          - string
          - 'null'
          format: uuid
        kind:
          $ref: '#/components/schemas/AlarmKind'
        label:
          type: string
        rhythm:
          $ref: '#/components/schemas/Rhythm'
        schedule:
          oneOf:
          - $ref: '#/components/schemas/Schedule'
          - type: 'null'
      additionalProperties: false
    AlarmKind:
      type: string
      enum:
      - scheduled
      - webhook
      - relay
    AlarmList:
      type: object
      required:
      - alarms
      properties:
        alarms:
          type: array
          items:
            $ref: '#/components/schemas/Alarm'
    AlarmPatch:
      type: object
      properties:
        channels:
          type:
          - array
          - 'null'
          items:
            $ref: '#/components/schemas/Channel'
        enabled:
          type:
          - boolean
          - 'null'
        kind:
          oneOf:
          - $ref: '#/components/schemas/AlarmKind'
          - type: 'null'
        label:
          type:
          - string
          - 'null'
        rhythm:
          oneOf:
          - $ref: '#/components/schemas/Rhythm'
          - type: 'null'
        schedule:
          oneOf:
          - $ref: '#/components/schemas/Schedule'
          - type: 'null'
      additionalProperties: false
...
    Channel:
      type: string
      enum:
      - phone
      - band
...
    DispatchIdResponse:
      type: object
      description: Returned by the test route and by webhook ingress.
      required:
      - dispatch_id
      properties:
        dispatch_id:
          type: string
          format: uuid
...
    NamedRhythm:
      type: string
      description: Built-in rhythm names. PLAN.md §9.2.
      enum:
      - single
      - double
      - triple
      - long
      - ramp
      - sos
...
    Problem:
      type: object
      description: RFC 9457 problem document. Handlers build it in `error.rs`; this type only documents it.
      required:
      - type
      - title
      - status
      - detail
      properties:
        current:
          type:
          - object
          - 'null'
          description: The stored entity, on `conflict` responses only.
        detail:
          type: string
        status:
          type: integer
          format: int32
          minimum: 0
        title:
          type: string
        type:
          type: string
          description: '`urn:icarus:problem:<slug>`, where the slug is one of the contract''s list.'
...
    Rhythm:
      oneOf:
      - $ref: '#/components/schemas/NamedRhythm'
      - type: array
        items:
          $ref: '#/components/schemas/Step'
      description: A built-in name, or an explicit list of steps.
...
    Schedule:
      type: object
      description: Scheduled alarms only. `weekdays` uses ISO numbering (1 = Monday). Empty means the next occurrence.
      required:
      - time
      - weekdays
      properties:
        time:
          type: string
        weekdays:
          type: array
          items:
            type: integer
            format: int32
            minimum: 0
...
    Step:
      oneOf:
      - type: object
        required:
        - preset
        - loops
        - type
        properties:
          loops:
            type: integer
            format: int32
            minimum: 0
          preset:
            type: integer
            format: int32
            minimum: 0
          type:
            type: string
            enum:
            - buzz
      - type: object
        required:
        - ms
        - type
        properties:
          ms:
            type: integer
            format: int32
            minimum: 0
          type:
            type: string
            enum:
            - pause
      description: One rhythm step. A buzz loop counts as 1 s toward the 30 s limit.
...
  securitySchemes:
    device:
      type: http
      scheme: bearer
    session:
      type: apiKey
      in: cookie
      name: icarus_session
      description: 'Website session from POST /v1/auth/login. Non-GET requests also send X-Icarus-CSRF: 1.'
...